refactor: reused bird utils constants
