+++ server/utilities/constants/preprocess/prepare_sample_dataset/indexing_constants.py
@@ -3,8 +3,5 @@
 This module defines key constants that are used to access and reference specific fields
 in the dataset during preprocessing and preparation of Text-to-SQL sample datasets.
 """
-
-QUESTION_ID="question_id"
-DB_ID="db_id"
 SCHEMA_USED = "schema_used"
 SQL="SQL"
